// src/apis/models/Volo_Abp_OpenIddict_Applications_Dtos_ApplicationTokenLifetimeDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Abp_OpenIddict_Applications_Dtos_ApplicationTokenLifetimeDto = {
    accessTokenLifetime?: number | null;
    authorizationCodeLifetime?: number | null;
    deviceCodeLifetime?: number | null;
    identityTokenLifetime?: number | null;
    refreshTokenLifetime?: number | null;
    userCodeLifetime?: number | null;
};